Migrate cookie constants usages to PSR
