+++ src/Adapter/Preferences/PreferencesConfiguration.php
@@ -26,9 +26,9 @@
 
 namespace PrestaShop\PrestaShop\Adapter\Preferences;
 
-use Cookie;
 use PrestaShop\PrestaShop\Adapter\Configuration;
 use PrestaShop\PrestaShop\Core\Configuration\DataConfigurationInterface;
+use PrestaShop\PrestaShop\Core\Http\CookieOptions;
 
 /**
  * This class will provide Shop Preferences configuration.
@@ -124,7 +124,7 @@ class PreferencesConfiguration implements DataConfigurationInterface
             $configuration['enable_ssl'] === false
             || $configuration['enable_ssl_everywhere'] === false
         )
-            && $this->configuration->get('PS_COOKIE_SAMESITE') === Cookie::SAMESITE_NONE;
+            && $this->configuration->get('PS_COOKIE_SAMESITE') === CookieOptions::SAMESITE_NONE;
     }
 
     /**
