+++ src/PrestaShopBundle/Form/Admin/Configure/AdvancedParameters/Administration/GeneralDataProvider.php
@@ -28,9 +28,9 @@ declare(strict_types=1);
 
 namespace PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Administration;
 
-use Cookie;
 use PrestaShop\PrestaShop\Core\Configuration\DataConfigurationInterface;
 use PrestaShop\PrestaShop\Core\Form\FormDataProviderInterface;
+use PrestaShop\PrestaShop\Core\Http\CookieOptions;
 use PrestaShopBundle\Form\Exception\DataProviderException;
 use PrestaShopBundle\Form\Exception\InvalidConfigurationDataError;
 use PrestaShopBundle\Form\Exception\InvalidConfigurationDataErrorCollection;
@@ -46,6 +46,8 @@ final class GeneralDataProvider implements FormDataProviderInterface
      * Hours are converted to seconds, so int might be turned to float if it's way to high.
      * Cookie classes crash if lifetime goes beyond year 9999, there are probably multiple other things.
      * So we need to set some sort of max value. 100 years seems like a lifetime beyond reasonable use.
+     *
+     * @deprecated since 9.0 use CookieOptions constants instead.
      */
     public const MAX_COOKIE_VALUE = 876000;
 
@@ -106,7 +108,7 @@ final class GeneralDataProvider implements FormDataProviderInterface
                 $errors->add(new InvalidConfigurationDataError(FormDataProvider::ERROR_NOT_NUMERIC_OR_LOWER_THAN_ZERO, GeneralType::FIELD_FRONT_COOKIE_LIFETIME));
             }
 
-            if ($frontOfficeLifeTimeCookie > self::MAX_COOKIE_VALUE) {
+            if ($frontOfficeLifeTimeCookie > CookieOptions::MAX_COOKIE_VALUE) {
                 $errors->add(new InvalidConfigurationDataError(FormDataProvider::ERROR_COOKIE_LIFETIME_MAX_VALUE_EXCEEDED, GeneralType::FIELD_FRONT_COOKIE_LIFETIME));
             }
         }
@@ -117,7 +119,7 @@ final class GeneralDataProvider implements FormDataProviderInterface
                 $errors->add(new InvalidConfigurationDataError(FormDataProvider::ERROR_NOT_NUMERIC_OR_LOWER_THAN_ZERO, GeneralType::FIELD_BACK_COOKIE_LIFETIME));
             }
 
-            if ($backOfficeLifeTimeCookie > self::MAX_COOKIE_VALUE) {
+            if ($backOfficeLifeTimeCookie > CookieOptions::MAX_COOKIE_VALUE) {
                 $errors->add(new InvalidConfigurationDataError(FormDataProvider::ERROR_COOKIE_LIFETIME_MAX_VALUE_EXCEEDED, GeneralType::FIELD_BACK_COOKIE_LIFETIME));
             }
         }
@@ -143,7 +145,7 @@ final class GeneralDataProvider implements FormDataProviderInterface
      */
     protected function validateSameSite(string $sameSite): bool
     {
-        if ($sameSite === Cookie::SAMESITE_NONE) {
+        if ($sameSite === CookieOptions::SAMESITE_NONE) {
             return $this->sslEnabled && $this->sslEnabledEverywhere;
         }
